Move `projectStorage` to `services` directory
